Moved jschema components under jschema folder
