+++ components/__tests__/jschema/data/ChannelInvalidJsonData.json
