+++ components/__tests__/jschema/data/ChannelInvalidJsonSchema.json
