+++ components/__tests__/jschema/data/ChannelJsonSchema.json
