+++ components/__tests__/jschema/properties/allOf.test.js
