+++ components/__tests__/jschema/properties/boolean.test.js
