+++ components/src/lib/jschema/properties/BooleanProperty.svelte
