+++ components/src/lib/jschema/properties/CollapsibleProperty.svelte
