+++ components/src/lib/jschema/properties/NumberProperty.svelte
