+++ components/src/lib/jschema/properties/PropertyDescription.svelte
