+++ components/src/lib/jschema/properties/PropertyLabel.svelte
