+++ components/src/lib/jschema/property_utils.js
