+++ src/lib/components/v1/workflow/ArgumentsSchema.svelte
@@ -6,7 +6,7 @@
 	import ImportExportArgs from './ImportExportArgs.svelte';
 	import { JSchema, getPropertiesToIgnore } from 'fractal-components';
 	import { deepCopy } from '$lib/common/component_utilities';
-	import { JsonSchemaDataError } from 'fractal-components/components/form_manager';
+	import { JsonSchemaDataError } from 'fractal-components/jschema/form_manager';
 
 	const SUPPORTED_SCHEMA_VERSIONS = ['pydantic_v1', 'pydantic_v2'];
