+++ src/lib/components/v2/workflow/ArgumentsSchema.svelte
@@ -15,7 +15,7 @@
 	import FormBuilder from './FormBuilder.svelte';
 	import { deepCopy } from '$lib/common/component_utilities';
 	import { tick } from 'svelte';
-	import { JsonSchemaDataError } from 'fractal-components/components/form_manager';
+	import { JsonSchemaDataError } from 'fractal-components/jschema/form_manager';
 
 	const SUPPORTED_SCHEMA_VERSIONS = ['pydantic_v1', 'pydantic_v2'];
